src/proxy/python/echo_client/pyproject.toml
[project]
name = "echo-client"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.68.1",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.68.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app/echo_client"]
sources = ["app"]

[tool.hatch.version]
path = "VERSION"

